* Update build system for SCons v1 and newer. * Skip filling PGM SKB or resolving CMSGs when recv() fails.

+++ trunk/openpgm/pgm/include/pgm/skbuff.h
@@ -227,7 +227,7 @@ static inline gboolean pgm_skb_is_valid (const struct pgm_sk_buff_t* const skb)
 	}
 /* truesize */
 	g_return_val_if_fail (skb->truesize >= sizeof(struct pgm_sk_buff_t*) + skb->len, FALSE);
-	g_return_val_if_fail (skb->truesize == (const guint8*)skb->end - (const guint8*)skb, FALSE);
+	g_return_val_if_fail (skb->truesize == (guint)((const guint8*)skb->end - (const guint8*)skb), FALSE);
 /* users */
 	g_return_val_if_fail (g_atomic_int_get (&skb->users) > 0, FALSE);
 	return TRUE;
